// src/generated/openapi/models/DerivedVusSummaryDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VIPLEV
 * This is an API specification
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Aggregated virtual user statistics for the run
 * @export
 * @interface DerivedVusSummaryDTO
 */
export interface DerivedVusSummaryDTO {
    /**
     * 
     * @type {number}
     * @memberof DerivedVusSummaryDTO
     */
    min?: number;
    /**
     * 
     * @type {number}
     * @memberof DerivedVusSummaryDTO
     */
    max?: number;
    /**
     * 
     * @type {number}
     * @memberof DerivedVusSummaryDTO
     */
    avg?: number;
}

/**
 * Check if a given object implements the DerivedVusSummaryDTO interface.
 */
export function instanceOfDerivedVusSummaryDTO(value: object): value is DerivedVusSummaryDTO {
    return true;
}

export function DerivedVusSummaryDTOFromJSON(json: any): DerivedVusSummaryDTO {
    return DerivedVusSummaryDTOFromJSONTyped(json, false);
}

export function DerivedVusSummaryDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DerivedVusSummaryDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'min': json['min'] == null ? undefined : json['min'],
        'max': json['max'] == null ? undefined : json['max'],
        'avg': json['avg'] == null ? undefined : json['avg'],
    };
}

export function DerivedVusSummaryDTOToJSON(json: any): DerivedVusSummaryDTO {
    return DerivedVusSummaryDTOToJSONTyped(json, false);
}

export function DerivedVusSummaryDTOToJSONTyped(value?: DerivedVusSummaryDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'min': value['min'],
        'max': value['max'],
        'avg': value['avg'],
    };
}